FPS Alpha .03
Added the source files for my project in zip format. This version has none of the nice textures as they're too big and I'm working on getting them scaled down.

This version is feature complete minus the moving platforms and nicer textures as aforementioned. 

Also, here's a link to the game as an executable if you're interested:
http://www.4shared.com/archive/5g0izWoOce/FPS03.html?